refactor: split main.go (1723 lines) into per-command files
main.go was a god file holding the CLI dispatch, the Bubble Tea
lifecycle, and every command body inline. After the deepening landed
across #249 (Client), #250 (deadline/time), #247 (Urgency), and #248
(goaltable), each command body shrank to a few flag-parse + filter +
render lines — so splitting them out by command is now a pure
relocation with no behaviour change.

main.go drops to 188 lines: just main(), the dispatch switch, printHelp,
printVersion, parseNoColorFlag, and loadConfigAndGoals (shared between
next and schedule).

New files, named to match the existing per-command test files
(next_test.go, list_test.go, view_test.go, schedule_test.go) where they
exist:

  - tui.go              Bubble Tea Init/Update/View lifecycle on `model`
  - next.go             buzz next + watch mode + clearScreen helpers
  - list.go             buzz list + getDisplayUnits + formatListRate
  - filtered.go         buzz all/today/tomorrow/due/less + filter
                        wrappers + tomorrow-view baremin helpers
  - add.go              buzz add + printAddUsageAndExit
  - refresh.go          buzz refresh
  - view.go             buzz view (web/json/text)
  - charge.go           buzz charge
  - deadline_cmd.go     buzz deadline (avoids clash with deadline.go
                        which holds the time-domain predicates)
  - uncle.go            buzz uncle
  - schedule.go         buzz schedule + timeSlot + density/timeline
                        renderers
  - review.go (mod)     gained handleReviewCommand alongside the
                        existing review-TUI Bubble Tea code

Net: main.go −1535 lines; new files +1545 lines (identical bodies, plus
the per-file package/import boilerplate Go requires).

Closes #244
