+++ main.go
@@ -1,234 +1,22 @@
 package main
 
 import (
-	"bytes"
-	"encoding/json"
-	"errors"
-	"flag"
 	"fmt"
-	"math"
 	"os"
-	"os/signal"
-	"sort"
-	"strconv"
-	"strings"
-	"syscall"
 	"time"
 
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
-	"github.com/charmbracelet/x/term"
 	"github.com/muesli/termenv"
 )
 
 // version is set via ldflags during build
 var version = "dev"
 
-// navigationTimeout is the duration of inactivity before the cell highlight is auto-disabled
-const navigationTimeout = 3 * time.Second
-
-// limsumFetchDelay is the duration to wait after adding a datapoint before fetching the updated limsum
-// This gives the Beeminder server time to update the goal's limsum with the new datapoint
-// This is a variable (not const) to allow tests to override it
+// limsumFetchDelay is the duration to wait after adding a datapoint before fetching the updated limsum.
+// This gives the Beeminder server time to update the goal's limsum with the new datapoint.
+// This is a variable (not const) to allow tests to override it.
 var limsumFetchDelay = 2 * time.Second
-
-func (m model) Init() tea.Cmd {
-	if m.state == "auth" {
-		return m.authModel.Init()
-	}
-	// In app state, load goals and start refresh timer
-	return tea.Batch(
-		loadGoalsCmd(m.appModel.client),
-		refreshTickCmd(),
-		checkRefreshFlagCmd(),
-	)
-}
-
-func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	// Handle window size messages for both states
-	if msg, ok := msg.(tea.WindowSizeMsg); ok {
-		m.width = msg.Width
-		m.height = msg.Height
-		m.appModel.width = msg.Width
-		m.appModel.height = msg.Height
-		// Re-clamp scroll position to keep cursor visible after resize
-		if m.state == "app" {
-			displayGoals := m.appModel.getDisplayGoals()
-			updateScrollForCursor(&m, len(displayGoals))
-		}
-	}
-
-	if m.state == "auth" {
-		// Handle auth state
-		switch msg := msg.(type) {
-		case authSuccessMsg:
-			// Authentication succeeded, switch to app
-			m.state = "app"
-			m.appModel = initialAppModel(msg.config)
-			m.appModel.width = m.width
-			m.appModel.height = m.height
-			return m, loadGoalsCmd(m.appModel.client)
-		default:
-			var cmd tea.Cmd
-			updatedModel, cmd := m.authModel.Update(msg)
-			if authModel, ok := updatedModel.(authModel); ok {
-				m.authModel = authModel
-			} else {
-				// Type assertion failed - log error and keep current authModel unchanged
-				fmt.Fprintf(os.Stderr, "Warning: authModel.Update returned unexpected type %T, keeping current authModel\n", updatedModel)
-				cmd = nil // Return safe command
-			}
-			return m, cmd
-		}
-	}
-
-	// Handle app state
-	return m.updateApp(msg)
-}
-
-func (m model) updateApp(msg tea.Msg) (tea.Model, tea.Cmd) {
-	switch msg := msg.(type) {
-
-	case goalsLoadedMsg:
-		// Goals have been loaded from the API
-		m.appModel.loading = false
-		if msg.err != nil {
-			m.appModel.err = msg.err
-		} else {
-			m.appModel.goals = msg.goals
-			m.appModel.err = nil
-		}
-		return m, nil
-
-	case refreshTickMsg:
-		// Time to refresh data
-		if m.appModel.refreshActive {
-			return m, tea.Batch(
-				loadGoalsCmd(m.appModel.client),
-				refreshTickCmd(), // Schedule the next refresh
-			)
-		}
-		return m, nil
-
-	case datapointSubmittedMsg:
-		// Datapoint submission completed
-		m.appModel.submitting = false
-		if msg.err != nil {
-			m.appModel.inputError = fmt.Sprintf("Failed to submit: %v", msg.err)
-		} else {
-			// Success - exit input mode and refresh goals (without showing loading state)
-			m.appModel.inputMode = false
-			m.appModel.inputFocus = 0
-			m.appModel.inputError = ""
-			// Don't set loading = true here to avoid the full-app loading state
-			return m, loadGoalsCmd(m.appModel.client)
-		}
-		return m, nil
-
-	case goalDetailsLoadedMsg:
-		// Goal details with datapoints have been loaded
-		if msg.err != nil {
-			// Error loading goal details - continue with basic goal info
-			return m, nil
-		}
-		if m.appModel.showModal && m.appModel.modalGoal != nil && msg.goal != nil {
-			// Update the modal goal with the detailed information
-			if m.appModel.modalGoal.Slug == msg.goal.Slug {
-				m.appModel.modalGoal = msg.goal
-			}
-		}
-		return m, nil
-
-	case goalCreatedMsg:
-		// Goal creation completed
-		m.appModel.creatingGoal = false
-		if msg.err != nil {
-			m.appModel.createError = fmt.Sprintf("Failed to create goal: %v", msg.err)
-		} else {
-			// Success - close modal and refresh goals
-			m.appModel.showCreateModal = false
-			m.appModel.createError = ""
-			return m, loadGoalsCmd(m.appModel.client)
-		}
-		return m, nil
-
-	case checkRefreshFlagMsg:
-		// Check if another process requested a refresh
-		flagTimestamp := getRefreshFlagTimestamp()
-		if flagTimestamp > m.lastRefreshTimestamp {
-			// New refresh event detected - update our last processed timestamp
-			m.lastRefreshTimestamp = flagTimestamp
-			return m, tea.Batch(
-				loadGoalsCmd(m.appModel.client),
-				checkRefreshFlagCmd(), // Schedule next check
-			)
-		}
-		// No new refresh event, but continue checking
-		return m, checkRefreshFlagCmd()
-
-	case navigationTimeoutMsg:
-		// Auto-disable highlight after inactivity
-		// Only disable if not in modal or search mode
-		if !m.appModel.showModal && !m.appModel.searchMode {
-			// Check if enough time has elapsed since last navigation
-			elapsed := time.Since(m.appModel.lastNavigationTime)
-			if elapsed >= navigationTimeout {
-				m.appModel.hasNavigated = false
-			}
-		}
-		return m, nil
-
-	// Is it a key press?
-	case tea.KeyMsg:
-		return handleKeyPress(m, msg)
-	}
-
-	// Return the updated model to the Bubble Tea runtime for processing.
-	// Note that we're not returning a command.
-	return m, nil
-}
-
-func (m model) View() string {
-	if m.state == "auth" {
-		return m.authModel.View()
-	}
-	return m.viewApp()
-}
-
-func (m model) viewApp() string {
-	if m.appModel.loading {
-		return "Loading goals...\n\nPress q to quit.\n"
-	}
-
-	if m.appModel.err != nil {
-		return fmt.Sprintf("Error loading goals: %v\n\nPress q to quit.\n", m.appModel.err)
-	}
-
-	// Get the goals to display (filtered or all)
-	displayGoals := m.appModel.getDisplayGoals()
-
-	// Render the grid and footer
-	grid := RenderGrid(displayGoals, m.appModel.width, m.appModel.height, m.appModel.scrollRow, m.appModel.cursor, m.appModel.hasNavigated, m.appModel.config.Username, m.appModel.searchMode, m.appModel.searchQuery)
-	footer := RenderFooter(displayGoals, m.appModel.width, m.appModel.height, m.appModel.scrollRow, m.appModel.refreshActive)
-
-	baseView := grid + footer
-
-	// Show create goal modal if active
-	if m.appModel.showCreateModal {
-		modal := RenderCreateGoalModal(m.appModel.width, m.appModel.height, m.appModel.createSlug, m.appModel.createTitle,
-			m.appModel.createGoalType, m.appModel.createGunits, m.appModel.createGoaldate, m.appModel.createGoalval,
-			m.appModel.createRate, m.appModel.createFocus, m.appModel.createError, m.appModel.creatingGoal)
-		return modal
-	}
-
-	// Show modal overlay if modal is active
-	if m.appModel.showModal && m.appModel.modalGoal != nil {
-		modal := RenderModal(m.appModel.modalGoal, m.appModel.width, m.appModel.height, m.appModel.inputDate, m.appModel.inputValue, m.appModel.inputComment, m.appModel.inputFocus, m.appModel.inputMode, m.appModel.inputError, m.appModel.submitting)
-		return modal
-	}
-
-	return baseView
-}
 
 func printHelp() {
 	fmt.Println("buzz - A terminal user interface for Beeminder")
@@ -374,45 +162,10 @@ func main() {
 	}
 }
 
-// handleNextCommand outputs a terse summary of the next due goal
-func handleNextCommand() {
-	// Parse flags for the next command
-	nextFlags := flag.NewFlagSet("next", flag.ContinueOnError)
-	watch := nextFlags.Bool("watch", false, "Watch mode - continuously refresh every 5 minutes")
-	watchShort := nextFlags.Bool("w", false, "Watch mode - continuously refresh every 5 minutes (shorthand)")
-	if err := nextFlags.Parse(os.Args[2:]); err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			// Help was requested; print usage and exit 0
-			fmt.Println("Usage: buzz next [-w|--watch]")
-			return
-		}
-		fmt.Fprintf(os.Stderr, "Error parsing flags: %s\n", redactError(err))
-		fmt.Fprintln(os.Stderr, "Usage: buzz next [-w|--watch]")
-		os.Exit(2)
-	}
-	if args := nextFlags.Args(); len(args) > 0 {
-		fmt.Fprintf(os.Stderr, "Unknown arguments: %v\n", args)
-		fmt.Fprintln(os.Stderr, "Usage: buzz next [-w|--watch]")
-		os.Exit(2)
-	}
-
-	// If either watch flag is set, enable watch mode
-	watchMode := *watch || *watchShort
-
-	if watchMode {
-		runWatchMode()
-	} else {
-		// One-shot mode - display and exit
-		if err := displayNextGoal(); err != nil {
-			fmt.Fprintf(os.Stderr, "Error: %s\n", redactError(err))
-			os.Exit(1)
-		}
-	}
-}
-
 // loadConfigAndGoals loads configuration, constructs an HTTP client, and fetches
 // sorted goals from Beeminder. Returns the client so callers can make further API
-// calls without rebuilding it.
+// calls without rebuilding it. Lives here because both next and schedule
+// commands use it.
 func loadConfigAndGoals() (*Config, Client, []Goal, error) {
 	if !ConfigExists() {
 		return nil, nil, nil, fmt.Errorf("no configuration found. Please run 'buzz' first to authenticate")
@@ -433,1291 +186,3 @@ func loadConfigAndGoals() (*Config, Client, []Goal, error) {
 	return config, client, goals, nil
 }
 
-// displayNextGoal fetches and displays the next due goal
-// Returns error instead of calling os.Exit() for reusability in watch mode
-func displayNextGoal() error {
-	_, _, goals, err := loadConfigAndGoals()
-	if err != nil {
-		return err
-	}
-
-	// If no goals, return error
-	if len(goals) == 0 {
-		return fmt.Errorf("no goals found")
-	}
-
-	// Get the first goal (most urgent)
-	nextGoal := goals[0]
-
-	// Format the output: "goalslug baremin timeframe"
-	timeframe := FormatDueDate(nextGoal.Losedate)
-
-	// Output the terse summary
-	fmt.Printf("%s %s %s\n", nextGoal.Slug, nextGoal.Baremin, timeframe)
-
-	// Check for updates and display message if available
-	fmt.Print(getUpdateMessage())
-
-	return nil
-}
-
-// runWatchMode runs the next command in watch mode with periodic refresh
-func runWatchMode() {
-	ticker := time.NewTicker(RefreshInterval)
-	defer ticker.Stop()
-
-	// Signal handling for graceful shutdown
-	sigChan := make(chan os.Signal, 1)
-	signal.Notify(sigChan, os.Interrupt, syscall.SIGTERM)
-	defer signal.Stop(sigChan)
-
-	// Initial display
-	clearScreen()
-	displayNextGoalWithTimestamp()
-
-	for {
-		select {
-		case <-ticker.C:
-			clearScreen()
-			displayNextGoalWithTimestamp()
-		case <-sigChan:
-			fmt.Println("\nExiting...")
-			return
-		}
-	}
-}
-
-// clearScreen clears the terminal screen
-func clearScreen() {
-	if fi, err := os.Stdout.Stat(); err == nil && (fi.Mode()&os.ModeCharDevice) == 0 {
-		return // not a terminal; skip clearing
-	}
-	fmt.Print("\033[2J\033[H")
-}
-
-// displayNextGoalWithTimestamp displays the next goal with a timestamp and refresh info
-func displayNextGoalWithTimestamp() {
-	fmt.Printf("[%s]\n", time.Now().Format("2006-01-02 15:04:05"))
-	if err := displayNextGoal(); err != nil {
-		fmt.Fprintf(os.Stderr, "Error: %s\n", redactError(err))
-	}
-	fmt.Printf("\nRefreshing every %dm... (Press Ctrl+C to exit)\n", int(RefreshInterval.Minutes()))
-}
-
-// bareminByEndOfTomorrowAt returns the baremin string to display for a goal in
-// the "due tomorrow" view. For goals already due tomorrow, the goal's existing
-// Baremin string is returned (it already reflects what's needed by tomorrow's
-// deadline). For goals due today, one day's worth of rate is added so the
-// displayed amount reflects what's required to avoid a beemergency tomorrow.
-//
-// The per-day slope is taken from the bright-line segment containing `now`
-// (via roadall) rather than from g.Rate, because g.Rate reports the goal's
-// end-of-graph rate which can differ from the current segment's rate for
-// goals with piecewise schedules.
-//
-// Two baremin value shapes are recognised: plain numeric (e.g. "+2") and the
-// colon-separated time format used by hour-valued goals — both "HH:MM" (e.g.
-// "+00:25") and "HH:MM:SS" (e.g. "+00:25:00"). For both, the slope is
-// interpreted as units-of-baremin per day before being added. The output
-// preserves whichever colon format the input used. If the slope can't be
-// determined or the value can't be parsed, the original Baremin string is
-// returned with any trailing time-window qualifier removed. The returned
-// string never carries that qualifier (e.g. " in 1 day", " within 1 day",
-// " today") — every row in the tomorrow view shares the same horizon, so
-// the suffix is just noise.
-func bareminByEndOfTomorrowAt(g Goal, now time.Time) string {
-	if !dueLaterTodayAt(g, now) {
-		return stripTimeWindowSuffix(g.Baremin)
-	}
-	if bumped, ok := bareminFromDueby(g, now); ok {
-		return stripTimeWindowSuffix(bumped)
-	}
-	perDay, ok := slopePerDayAt(g, now)
-	if !ok {
-		return stripTimeWindowSuffix(g.Baremin)
-	}
-	value := ParseBareminValue(g.Baremin)
-
-	// Colon-separated time formats — HH:MM or HH:MM:SS. The base value and
-	// the rate share the same hour unit, so add in seconds and reformat in
-	// whichever colon format the input used.
-	if strings.Contains(value, ":") {
-		baseSeconds, includeSeconds, ok := parseTimeValue(value)
-		if !ok {
-			return stripTimeWindowSuffix(g.Baremin)
-		}
-		totalSeconds := baseSeconds + int(math.Round(perDay*3600))
-		return formatTimeValue(totalSeconds, includeSeconds)
-	}
-
-	// Plain numeric.
-	base, err := strconv.ParseFloat(value, 64)
-	if err != nil {
-		return stripTimeWindowSuffix(g.Baremin)
-	}
-	total := base + perDay
-	sign := "+"
-	if total < 0 {
-		sign = ""
-	}
-	return fmt.Sprintf("%s%g", sign, total)
-}
-
-// stripTimeWindowSuffix removes Beeminder's trailing time-window phrase from a
-// baremin/limsum string (e.g. " in 1 day", " within 1 day", " in 3 hours",
-// " today"), leaving just the leading signed value. Used by the tomorrow view,
-// where every row shares the same horizon so the suffix only adds noise.
-func stripTimeWindowSuffix(s string) string {
-	if s == "" {
-		return s
-	}
-	for _, sep := range []string{" within ", " in "} {
-		if i := strings.LastIndex(s, sep); i >= 0 {
-			return s[:i]
-		}
-	}
-	// Handle "+0 today" / "0 today" style: drop the trailing " today" word.
-	if strings.HasSuffix(s, " today") {
-		return strings.TrimSuffix(s, " today")
-	}
-	return s
-}
-
-// bareminFromDueby returns the formatted delta for tomorrow's daystamp,
-// taken from Beeminder's pre-rounded `dueby` map. Beeminder rounds those
-// strings to the goal's Display Precision, so honouring them sidesteps
-// float-formatting noise (e.g. "+10788.140000000001") that arises when we
-// add today's baremin to a per-day rate ourselves. Returns ok=false when
-// dueby is absent or the tomorrow entry is missing/empty — callers fall
-// back to local arithmetic in those cases.
-func bareminFromDueby(g Goal, now time.Time) (string, bool) {
-	if len(g.Dueby) == 0 {
-		return "", false
-	}
-	entry, ok := g.Dueby[tomorrowDaystampFor(g, now)]
-	if !ok || entry.FormattedDelta == "" {
-		return "", false
-	}
-	return entry.FormattedDelta, true
-}
-
-// tomorrowDaystampFor returns the YYYYMMDD daystamp representing the day
-// after the goal's current Beeminder day. Beeminder shifts day boundaries
-// by the goal's `deadline` seconds — positive deadlines push the boundary
-// past midnight (e.g. +10800 = 3am cutoff), negative deadlines pull it
-// before (e.g. -10800 = 9pm cutoff). Subtracting the deadline from `now`
-// produces a wall-clock time inside the user's calendar date that matches
-// the goal's current daystamp; the next calendar day is tomorrow's.
-func tomorrowDaystampFor(g Goal, now time.Time) string {
-	shifted := now.Add(-time.Duration(g.Deadline) * time.Second).In(now.Location())
-	tomorrow := time.Date(shifted.Year(), shifted.Month(), shifted.Day(), 0, 0, 0, 0, now.Location()).AddDate(0, 0, 1)
-	return tomorrow.Format("20060102")
-}
-
-// isDoLessFilter returns true if the goal is a do-less type goal
-func isDoLessFilter(g Goal) bool {
-	return IsDoLessGoal(g)
-}
-
-// allGoalsFilter returns true for all goals
-func allGoalsFilter(g Goal) bool {
-	return true
-}
-
-// handleListCommand outputs a summary list of all goals with slug, title, rate, and stakes
-func handleListCommand() {
-	// Load config
-	if !ConfigExists() {
-		fmt.Println("Error: No configuration found. Please run 'buzz' first to authenticate.")
-		os.Exit(1)
-	}
-
-	config, err := LoadConfig()
-	if err != nil {
-		fmt.Printf("Error: Failed to load config: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	client := NewHTTPClient(config)
-
-	// Fetch goals
-	goals, err := client.FetchGoals()
-	if err != nil {
-		fmt.Printf("Error: Failed to fetch goals: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	// Sort goals alphabetically by slug for easy scanning
-	SortGoalsBySlug(goals)
-
-	// If no goals, exit
-	if len(goals) == 0 {
-		fmt.Println("No goals found.")
-		return
-	}
-
-	// Print summary header
-	fmt.Printf("Total goals: %d\n\n", len(goals))
-
-	table := Table{
-		ShowHeader: true,
-		Columns: []Column{
-			{Header: "Slug", Cell: func(g Goal) string { return g.Slug }},
-			{Header: "Title", Cell: func(g Goal) string {
-				if g.Title == "" {
-					return "-"
-				}
-				return g.Title
-			}},
-			{Header: "Units", Cell: func(g Goal) string { return getDisplayUnits(g.Gunits) }},
-			{Header: "Rate", Cell: func(g Goal) string { return formatListRate(g.Rate, g.Runits) }},
-			{Header: "Stakes", Cell: func(g Goal) string { return fmt.Sprintf("$%.2f", g.Pledge) }},
-		},
-	}
-	fmt.Print(table.Render(goals))
-
-	// Check for updates and display message if available
-	fmt.Print(getUpdateMessage())
-}
-
-// getDisplayUnits returns the display value for goal units, using "-" if empty
-func getDisplayUnits(gunits string) string {
-	if gunits == "" {
-		return "-"
-	}
-	return gunits
-}
-
-// formatListRate formats the rate value with its units for the list command
-func formatListRate(rate *float64, runits string) string {
-	if rate == nil {
-		return "-"
-	}
-	// Format rate to remove unnecessary decimal places
-	rateVal := *rate
-	if rateVal == float64(int(rateVal)) {
-		// Integer value - no decimal places
-		return fmt.Sprintf("%d/%s", int(rateVal), runits)
-	}
-	// Has decimal - use %.6g to show up to 6 significant digits, trimming trailing zeros
-	return fmt.Sprintf("%.6g/%s", rateVal, runits)
-}
-
-// handleAllCommand outputs all goals
-func handleAllCommand() {
-	handleFilteredCommand("all", allGoalsFilter)
-}
-
-// handleTodayCommand outputs all goals that are due today
-func handleTodayCommand() {
-	handleFilteredCommand("today", isDueTodayFilter)
-}
-
-// handleTomorrowCommand outputs all goals that are due tomorrow. Goals that
-// are already due today are included with their baremin bumped by one day's
-// rate, so the user sees the total amount they would need to do for the goal
-// to not be due tomorrow. The displayed deadline is bumped to match — the
-// bumped baremin is what's needed by *tomorrow's* deadline, not today's.
-func handleTomorrowCommand() {
-	now := time.Now()
-	filter := func(g Goal) bool { return isDueTomorrowFilterAt(g, now) }
-	bareminFor := func(g Goal) string { return bareminByEndOfTomorrowAt(g, now) }
-	losedateFor := func(g Goal) int64 { return losedateByEndOfTomorrowAt(g, now) }
-	handleFilteredCommandWithDisplay("tomorrow", filter, bareminFor, losedateFor)
-}
-
-// losedateByEndOfTomorrowAt returns the deadline timestamp to display for a
-// goal in the tomorrow view. For goals due later today (whose baremin we're
-// bumping by one day's rate), advance the deadline by one calendar day in
-// the caller's local zone so the displayed wall-clock deadline stays correct
-// across DST transitions. Overdue goals (losedate already in the past) keep
-// their own losedate so the OVERDUE indicator remains visible — bumping them
-// would silently hide the fact that they've already derailed.
-func losedateByEndOfTomorrowAt(g Goal, now time.Time) int64 {
-	if !dueLaterTodayAt(g, now) {
-		return g.Losedate
-	}
-	return time.Unix(g.Losedate, 0).In(now.Location()).AddDate(0, 0, 1).Unix()
-}
-
-// handleLessCommand outputs all do-less type goals
-func handleLessCommand() {
-	handleFilteredCommand("do-less", isDoLessFilter)
-}
-
-// handleDueCommand outputs all goals due within the specified duration
-func handleDueCommand() {
-	// Check arguments: buzz due <duration>
-	if len(os.Args) < 3 {
-		fmt.Println("Error: Missing required duration argument")
-		fmt.Println("Usage: buzz due <duration>")
-		fmt.Println("  Examples: buzz due 10m, buzz due 1h, buzz due 5d, buzz due 1w")
-		fmt.Println("  Supported units: m (minutes), h (hours), d (days), w (weeks)")
-		os.Exit(1)
-	}
-
-	durationStr := os.Args[2]
-
-	// Parse the duration
-	duration, ok := ParseDuration(durationStr)
-	if !ok {
-		fmt.Printf("Error: Invalid duration format: %s\n", durationStr)
-		fmt.Println("Usage: buzz due <duration>")
-		fmt.Println("  Examples: buzz due 10m, buzz due 1h, buzz due 5d, buzz due 1w")
-		fmt.Println("  Supported units: m (minutes), h (hours), d (days), w (weeks)")
-		os.Exit(1)
-	}
-
-	// Create filter function that captures the duration
-	isDueWithinFilter := func(g Goal) bool {
-		return IsDueWithin(g.Losedate, duration)
-	}
-
-	// Format the filter name for display
-	filterName := fmt.Sprintf("due within %s", durationStr)
-	handleFilteredCommand(filterName, isDueWithinFilter)
-}
-
-// handleFilteredCommand is a shared helper that outputs all goals matching the given filter
-// filterName is used in messages (e.g., "today", "tomorrow", or "do-less")
-// filter is a function that takes a Goal and returns true if the goal matches
-func handleFilteredCommand(filterName string, filter func(Goal) bool) {
-	handleFilteredCommandWithDisplay(filterName, filter,
-		func(g Goal) string { return g.Baremin },
-		func(g Goal) int64 { return g.Losedate },
-	)
-}
-
-// sortGoalsByDisplayedLosedate reorders goals in place so the slice ends up
-// sorted by the timestamp that losedateFor would render. SliceStable preserves
-// the input order for ties so any prior sort (e.g. SortGoals's pledge/slug
-// tiebreakers) survives.
-func sortGoalsByDisplayedLosedate(goals []Goal, losedateFor func(Goal) int64) {
-	sort.SliceStable(goals, func(i, j int) bool {
-		return losedateFor(goals[i]) < losedateFor(goals[j])
-	})
-}
-
-// handleFilteredCommandWithDisplay is the most general filtered-output helper:
-// the caller can override both the displayed baremin string and the deadline
-// timestamp used for the timeframe/absolute-deadline columns. The tomorrow
-// view uses this to bump both for due-today goals so the bumped baremin and
-// the displayed deadline are aligned to the same target moment.
-func handleFilteredCommandWithDisplay(filterName string, filter func(Goal) bool, bareminFor func(Goal) string, losedateFor func(Goal) int64) {
-	// Load config
-	if !ConfigExists() {
-		fmt.Println("Error: No configuration found. Please run 'buzz' first to authenticate.")
-		os.Exit(1)
-	}
-
-	config, err := LoadConfig()
-	if err != nil {
-		fmt.Printf("Error: Failed to load config: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	client := NewHTTPClient(config)
-
-	// Fetch goals
-	goals, err := client.FetchGoals()
-	if err != nil {
-		fmt.Printf("Error: Failed to fetch goals: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	// Sort goals (by due date ascending, then by stakes descending, then by name)
-	SortGoals(goals)
-
-	// Filter goals that match the criteria
-	var filteredGoals []Goal
-	for _, goal := range goals {
-		if filter(goal) {
-			filteredGoals = append(filteredGoals, goal)
-		}
-	}
-
-	// If no matching goals, exit
-	if len(filteredGoals) == 0 {
-		fmt.Printf("No %s goals found.\n", filterName)
-		return
-	}
-
-	// SortGoals ordered by each goal's own losedate, but the tomorrow view may
-	// show a bumped losedate for due-today goals. Re-sort by the displayed
-	// losedate so the rendered order matches the deadline column. SliceStable
-	// preserves the SortGoals tiebreakers (pledge desc, slug asc) when
-	// displayed losedates are equal.
-	sortGoalsByDisplayedLosedate(filteredGoals, losedateFor)
-
-	table := Table{
-		Colorize: true,
-		Columns: []Column{
-			{Cell: func(g Goal) string { return g.Slug }},
-			{Cell: func(g Goal) string { return bareminFor(g) }},
-			{Cell: func(g Goal) string { return FormatDueDate(losedateFor(g)) }},
-			{Cell: func(g Goal) string { return FormatAbsoluteDeadline(losedateFor(g)) }},
-		},
-	}
-	fmt.Print(table.Render(filteredGoals))
-
-	// Check for updates and display message if available
-	fmt.Print(getUpdateMessage())
-}
-
-// printAddUsageAndExit prints the usage for buzz add command and exits with code 1
-func printAddUsageAndExit(errorMsg string) {
-	fmt.Println("Error: " + errorMsg)
-	fmt.Println("Usage: buzz add [--requestid=<id>] [--daystamp=<date>] <goalslug> <value> [comment]")
-	fmt.Println("       echo \"<value>\" | buzz add [--requestid=<id>] [--daystamp=<date>] <goalslug> [comment]")
-	fmt.Println("")
-	fmt.Println("Note: Flags must come BEFORE positional arguments.")
-	fmt.Println("      Example: buzz add --daystamp=20240115 goalslug value comment")
-	fmt.Println("      The --daystamp flag accepts dates in YYYYMMDD format.")
-	os.Exit(1)
-}
-
-// handleAddCommand adds a datapoint to a goal without opening the TUI
-func handleAddCommand() {
-	// Parse flags for the add command
-	addFlags := flag.NewFlagSet("add", flag.ContinueOnError)
-	requestid := addFlags.String("requestid", "", "Request ID for idempotency")
-	daystamp := addFlags.String("daystamp", "", "Date for the datapoint in YYYYMMDD format")
-	if err := addFlags.Parse(os.Args[2:]); err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			fmt.Println("Usage: buzz add [--requestid=<id>] [--daystamp=<date>] <goalslug> <value> [comment]")
-			fmt.Println("       echo \"<value>\" | buzz add [--requestid=<id>] [--daystamp=<date>] <goalslug> [comment]")
-			fmt.Println("")
-			fmt.Println("Note: Flags must come BEFORE positional arguments.")
-			fmt.Println("      The --daystamp flag accepts dates in YYYYMMDD format.")
-			return
-		}
-		fmt.Fprintf(os.Stderr, "Error parsing flags: %s\n", redactError(err))
-		printAddUsageAndExit("Invalid flags")
-	}
-
-	// Get remaining positional arguments after flag parsing
-	args := addFlags.Args()
-
-	// Detect if known flags appear after positional arguments and warn the user
-	if misplacedFlag := detectMisplacedFlag(args); misplacedFlag != "" {
-		fmt.Fprintf(os.Stderr, "Warning: Flag '%s' appears after positional arguments and will be treated as part of the comment.\n", misplacedFlag)
-		fmt.Fprintf(os.Stderr, "Flags must come BEFORE positional arguments to be recognized.\n")
-		fmt.Fprintf(os.Stderr, "Correct usage: buzz add [--requestid=ID] [--daystamp=DATE] goalslug value comment\n")
-		fmt.Fprintln(os.Stderr, "")
-	}
-
-	// Check arguments: buzz add <goalslug> <value> [comment]
-	// Value can also be piped via stdin: echo "123" | buzz add mygoal [comment]
-	if len(args) < 1 {
-		printAddUsageAndExit("Missing required arguments")
-	}
-
-	goalSlug := args[0]
-	var value string
-	var commentStartIndex int // Index where optional comment starts in args
-
-	// Try to read value from stdin first (for piped input)
-	stdinValue, err := readValueFromStdin()
-	if err == nil && stdinValue != "" {
-		// Value provided via stdin
-		value = stdinValue
-		commentStartIndex = 1 // Comment starts at index 1 when value is piped
-	} else if len(args) >= 2 {
-		// Value provided as argument
-		value = args[1]
-		commentStartIndex = 2 // Comment starts at index 2
-	} else {
-		printAddUsageAndExit("Missing required value argument")
-	}
-
-	// Optional comment - default to "Added via buzz" if not provided
-	comment := "Added via buzz"
-	if len(args) >= commentStartIndex+1 {
-		comment = strings.Join(args[commentStartIndex:], " ")
-	}
-
-	// Load config
-	if !ConfigExists() {
-		fmt.Println("Error: No configuration found. Please run 'buzz' first to authenticate.")
-		os.Exit(1)
-	}
-
-	config, err := LoadConfig()
-	if err != nil {
-		fmt.Printf("Error: Failed to load config: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	client := NewHTTPClient(config)
-
-	// Parse and validate daystamp if provided
-	var daystampForAPI string
-	if *daystamp != "" {
-		// Validate date format (YYYYMMDD)
-		_, err := time.Parse("20060102", *daystamp)
-		if err != nil {
-			fmt.Printf("Error: Invalid date format for --daystamp: %s (expected YYYYMMDD)\n", *daystamp)
-			os.Exit(1)
-		}
-		daystampForAPI = *daystamp
-	}
-
-	// Use current time as timestamp (only used if daystamp is not provided)
-	timestamp := strconv.FormatInt(time.Now().Unix(), 10)
-
-	// Convert time format to decimal hours if needed
-	if isTimeFormat(value) {
-		decimalValue, ok := timeToDecimalHours(value)
-		if !ok {
-			fmt.Printf("Error: Invalid time format: %s\n", value)
-			os.Exit(1)
-		}
-		value = fmt.Sprintf("%.6g", decimalValue)
-	}
-
-	// Validate value is a number
-	if _, err := strconv.ParseFloat(value, 64); err != nil {
-		fmt.Printf("Error: Value must be a valid number, got: %s\n", value)
-		os.Exit(1)
-	}
-
-	// Create the datapoint
-	err = client.CreateDatapointWithDaystamp(goalSlug, timestamp, daystampForAPI, value, comment, *requestid)
-	if err != nil {
-		fmt.Printf("Error: Failed to add datapoint: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	// Signal any running TUI instances to refresh
-	if err := createRefreshFlag(); err != nil {
-		// Don't fail the command if flag creation fails
-		fmt.Fprintf(os.Stderr, "Warning: Could not create refresh flag: %s\n", redactError(err))
-	}
-
-	successMsg := fmt.Sprintf("Successfully added datapoint to %s: value=%s, comment=\"%s\"", goalSlug, value, comment)
-	if *daystamp != "" {
-		successMsg += fmt.Sprintf(", daystamp=%s", *daystamp)
-	}
-	if *requestid != "" {
-		successMsg += fmt.Sprintf(", requestid=\"%s\"", *requestid)
-	}
-	fmt.Println(successMsg)
-
-	// Wait briefly before fetching limsum to allow the server to update
-	time.Sleep(limsumFetchDelay)
-
-	// Fetch the goal to display the updated limsum
-	goal, err := client.FetchGoal(goalSlug)
-	if err != nil {
-		// Don't fail the command if fetching limsum fails, just skip displaying it
-		fmt.Fprintf(os.Stderr, "Warning: Could not fetch goal status: %s\n", redactError(err))
-	} else {
-		fmt.Printf("Limsum: %s\n", goal.Limsum)
-	}
-
-	// Check for updates and display message if available
-	fmt.Print(getUpdateMessage())
-}
-
-// handleRefreshCommand refreshes autodata for a goal
-func handleRefreshCommand() {
-	// Check arguments: buzz refresh <goalslug>
-	if len(os.Args) < 3 {
-		fmt.Println("Error: Missing required argument")
-		fmt.Println("Usage: buzz refresh <goalslug>")
-		os.Exit(1)
-	}
-
-	goalSlug := os.Args[2]
-
-	// Load config
-	if !ConfigExists() {
-		fmt.Println("Error: No configuration found. Please run 'buzz' first to authenticate.")
-		os.Exit(1)
-	}
-
-	config, err := LoadConfig()
-	if err != nil {
-		fmt.Printf("Error: Failed to load config: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	client := NewHTTPClient(config)
-
-	// Refresh the goal
-	queued, err := client.RefreshGoal(goalSlug)
-	if err != nil {
-		fmt.Printf("Error: Failed to refresh goal: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	if queued {
-		fmt.Printf("Successfully queued refresh for goal: %s\n", goalSlug)
-	} else {
-		fmt.Printf("Goal %s was not queued for refresh\n", goalSlug)
-	}
-
-	// Check for updates and display message if available
-	fmt.Print(getUpdateMessage())
-}
-
-// handleViewCommand displays detailed information about a specific goal
-func handleViewCommand() {
-	// Parse flags for the view command
-	viewFlags := flag.NewFlagSet("view", flag.ContinueOnError)
-	web := viewFlags.Bool("web", false, "Open the goal in the browser")
-	jsonOutput := viewFlags.Bool("json", false, "Output goal data as JSON")
-	datapoints := viewFlags.Bool("datapoints", false, "Include datapoints in output (use with --json)")
-	if err := viewFlags.Parse(os.Args[2:]); err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			// Help was requested; print usage and exit 0
-			fmt.Println("Usage: buzz view <goalslug> [--web] [--json] [--datapoints]")
-			return
-		}
-		fmt.Fprintf(os.Stderr, "Error parsing flags: %s\n", redactError(err))
-		fmt.Fprintln(os.Stderr, "Usage: buzz view <goalslug> [--web] [--json] [--datapoints]")
-		os.Exit(2)
-	}
-
-	// Get goal slug from remaining arguments
-	args := viewFlags.Args()
-
-	// Check if flags appear after the goal slug (handle both positions)
-	webFlag := *web
-	jsonFlag := *jsonOutput
-	datapointsFlag := *datapoints
-	var goalSlug string
-	var filteredArgs []string
-
-	for _, arg := range args {
-		switch arg {
-		case "--web":
-			webFlag = true
-		case "--json":
-			jsonFlag = true
-		case "--datapoints":
-			datapointsFlag = true
-		default:
-			filteredArgs = append(filteredArgs, arg)
-		}
-	}
-
-	if len(filteredArgs) < 1 {
-		fmt.Fprintln(os.Stderr, "Error: Missing required argument")
-		fmt.Fprintln(os.Stderr, "Usage: buzz view <goalslug> [--web] [--json] [--datapoints]")
-		os.Exit(1)
-	}
-
-	goalSlug = filteredArgs[0]
-
-	// Load config
-	if !ConfigExists() {
-		fmt.Fprintln(os.Stderr, "Error: No configuration found. Please run 'buzz' first to authenticate.")
-		os.Exit(1)
-	}
-
-	config, err := LoadConfig()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: Failed to load config: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	client := NewHTTPClient(config)
-
-	// If --web flag is present, open in browser and exit
-	if webFlag {
-		if err := openBrowser(config, goalSlug); err != nil {
-			fmt.Fprintf(os.Stderr, "Error: Failed to open browser: %s\n", redactError(err))
-			os.Exit(1)
-		}
-		return
-	}
-
-	// If --json flag is present, fetch and output raw JSON
-	if jsonFlag {
-		rawJSON, err := client.FetchGoalRawJSON(goalSlug, datapointsFlag)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "Error: %s\n", redactError(err))
-			os.Exit(1)
-		}
-
-		// Pretty print the raw JSON
-		var prettyJSON bytes.Buffer
-		if err := json.Indent(&prettyJSON, rawJSON, "", "  "); err != nil {
-			fmt.Fprintf(os.Stderr, "Error: Failed to format JSON: %s\n", redactError(err))
-			os.Exit(1)
-		}
-		fmt.Println(prettyJSON.String())
-		return
-	}
-
-	// Warn if --datapoints is used without --json
-	if datapointsFlag {
-		fmt.Fprintln(os.Stderr, "Warning: --datapoints flag has no effect without --json")
-	}
-
-	// Fetch the goal for human-readable output
-	goal, err := client.FetchGoal(goalSlug)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	// Display goal information (human-readable format)
-	fmt.Printf("Goal: %s\n", goal.Slug)
-	fmt.Print(formatGoalDetails(goal, config))
-
-	// Check for updates and display message if available
-	fmt.Print(getUpdateMessage())
-}
-
-// handleReviewCommand launches an interactive review of all goals
-func handleReviewCommand() {
-	// Load config
-	if !ConfigExists() {
-		fmt.Fprintln(os.Stderr, "Error: No configuration found. Please run 'buzz' first to authenticate.")
-		os.Exit(1)
-	}
-
-	config, err := LoadConfig()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: Failed to load config: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	client := NewHTTPClient(config)
-
-	// Fetch goals
-	goals, err := client.FetchGoals()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: Failed to fetch goals: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	if len(goals) == 0 {
-		fmt.Println("No goals found.")
-		return
-	}
-
-	// Sort goals alphabetically by slug as specified
-	SortGoalsBySlug(goals)
-
-	// Launch the interactive review TUI
-	p := tea.NewProgram(initialReviewModel(goals, config), tea.WithAltScreen())
-	if _, err := p.Run(); err != nil {
-		fmt.Fprintf(os.Stderr, "Error: %s\n", redactError(err))
-		os.Exit(1)
-	}
-}
-
-// handleChargeCommand creates a charge for the authenticated user
-func handleChargeCommand() {
-	// Usage: buzz charge <amount> <note> [--dryrun]
-	if len(os.Args) < 4 {
-		fmt.Fprintln(os.Stderr, "Error: Missing required arguments")
-		fmt.Fprintln(os.Stderr, "Usage: buzz charge <amount> <note> [--dryrun]")
-		os.Exit(1)
-	}
-
-	args := os.Args[2:]
-	amountStr := args[0]
-	// Collect note parts and allow --dryrun anywhere after amount
-	dryrun := false
-	var noteParts []string
-	for _, a := range args[1:] {
-		if a == "--dryrun" {
-			dryrun = true
-			continue
-		}
-		noteParts = append(noteParts, a)
-	}
-	note := strings.Join(noteParts, " ")
-	if strings.TrimSpace(note) == "" {
-		fmt.Fprintln(os.Stderr, "Error: Note is required")
-		fmt.Fprintln(os.Stderr, "Usage: buzz charge <amount> <note> [--dryrun]")
-		os.Exit(1)
-	}
-
-	// Validate amount is a number
-	amount, err := strconv.ParseFloat(amountStr, 64)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: Amount must be a valid number, got: %s\n", amountStr)
-		os.Exit(1)
-	}
-
-	// Validate amount is >= 1.00
-	if amount < 1.00 {
-		fmt.Fprintf(os.Stderr, "Error: Amount must be at least 1.00, got: %.2f\n", amount)
-		os.Exit(1)
-	}
-
-	// Load config
-	if !ConfigExists() {
-		fmt.Fprintln(os.Stderr, "Error: No configuration found. Please run 'buzz' first to authenticate.")
-		os.Exit(1)
-	}
-
-	config, err := LoadConfig()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: Failed to load config: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	client := NewHTTPClient(config)
-
-	// Create the charge (API returns the created/dry-run charge)
-	ch, err := client.CreateCharge(amount, note, dryrun)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: Failed to create charge: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	if dryrun {
-		fmt.Printf("Dry run: Would charge $%.2f with note: %q for %s\n", ch.Amount, ch.Note, ch.Username)
-	} else {
-		fmt.Printf("Successfully created charge %s: $%.2f with note: %q for %s\n", ch.ID, ch.Amount, ch.Note, ch.Username)
-	}
-
-	// Check for updates and display message if available
-	fmt.Print(getUpdateMessage())
-}
-
-func handleDeadlineCommand() {
-	// Usage: buzz deadline [--yes] <goalslug> <time>
-	deadlineFlags := flag.NewFlagSet("deadline", flag.ContinueOnError)
-	yes := deadlineFlags.Bool("yes", false, "Skip confirmation prompt")
-	yesShort := deadlineFlags.Bool("y", false, "Skip confirmation prompt (shorthand)")
-	if err := deadlineFlags.Parse(os.Args[2:]); err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			fmt.Fprintln(os.Stderr, "Usage: buzz deadline [--yes|-y] <goalslug> <time>")
-			fmt.Fprintln(os.Stderr, "  <time> can be:")
-			fmt.Fprintln(os.Stderr, "    - 12-hour format: \"3:00 PM\", \"11:30 AM\"")
-			fmt.Fprintln(os.Stderr, "    - 24-hour format: \"15:00\", \"23:30\"")
-			return
-		}
-		fmt.Fprintf(os.Stderr, "Error parsing flags: %s\n", redactError(err))
-		fmt.Fprintln(os.Stderr, "Usage: buzz deadline [--yes|-y] <goalslug> <time>")
-		os.Exit(2)
-	}
-
-	args := deadlineFlags.Args()
-	if len(args) < 2 {
-		fmt.Fprintln(os.Stderr, "Error: Missing required arguments")
-		fmt.Fprintln(os.Stderr, "Usage: buzz deadline [--yes] <goalslug> <time>")
-		fmt.Fprintln(os.Stderr, "  <time> can be:")
-		fmt.Fprintln(os.Stderr, "    - 12-hour format: \"3:00 PM\", \"11:30 AM\"")
-		fmt.Fprintln(os.Stderr, "    - 24-hour format: \"15:00\", \"23:30\"")
-		os.Exit(1)
-	}
-
-	skipConfirm := *yes || *yesShort
-	goalSlug := args[0]
-	timeStr := strings.Join(args[1:], " ")
-
-	offset, err := parseTimeToDeadlineOffset(timeStr)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: %s\n", err)
-		os.Exit(1)
-	}
-
-	if !ConfigExists() {
-		fmt.Fprintln(os.Stderr, "Error: No configuration found. Please run 'buzz' first to authenticate.")
-		os.Exit(1)
-	}
-
-	config, err := LoadConfig()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: Failed to load config: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	client := NewHTTPClient(config)
-
-	// Fetch current goal to show existing deadline
-	currentGoal, err := client.FetchGoal(goalSlug)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: Failed to fetch goal: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	newTime := formatDueTime(offset)
-	currentTime := formatDueTime(currentGoal.Deadline)
-
-	if !skipConfirm {
-		fmt.Printf("Change deadline for %s from %s to %s? [y/N] ", goalSlug, currentTime, newTime)
-		var response string
-		fmt.Scanln(&response)
-		response = strings.TrimSpace(strings.ToLower(response))
-		if response != "y" && response != "yes" {
-			fmt.Println("Cancelled.")
-			return
-		}
-	}
-
-	goal, err := client.UpdateGoalDeadline(goalSlug, offset)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: Failed to update deadline: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	fmt.Printf("Updated deadline for %s to %s\n", goal.Slug, formatDueTime(goal.Deadline))
-
-	fmt.Print(getUpdateMessage())
-}
-
-// timeSlot represents a time of day with goals scheduled at that time
-type timeSlot struct {
-	hour   int      // 0-23
-	minute int      // 0-59
-	goals  []string // goal slugs at this time
-}
-
-// handleScheduleCommand displays a visual representation of goal deadline distribution throughout a 24-hour day
-func handleScheduleCommand() {
-	// Load config and goals
-	_, _, goals, err := loadConfigAndGoals()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	if len(goals) == 0 {
-		fmt.Println("No goals found.")
-		return
-	}
-
-	// Extract time-of-day from each goal's deadline
-	timeSlots := extractTimeSlots(goals)
-
-	// Generate hourly density data
-	hourCounts := make([]int, 24)
-	for _, slot := range timeSlots {
-		hourCounts[slot.hour] += len(slot.goals)
-	}
-
-	// Display hourly density overview
-	displayHourlyDensity(hourCounts)
-
-	// Display detailed timeline
-	displayTimeline(timeSlots)
-
-	// Check for updates and display message if available
-	fmt.Print(getUpdateMessage())
-}
-
-// extractTimeSlots extracts time-of-day from all goal deadlines and groups them
-func extractTimeSlots(goals []Goal) []timeSlot {
-	// Map to group goals by their time of day (hour:minute)
-	slotMap := make(map[string]*timeSlot)
-
-	for _, goal := range goals {
-		// Convert losedate to local time and extract hour/minute
-		t := time.Unix(goal.Losedate, 0).In(time.Local)
-		hour := t.Hour()
-		minute := t.Minute()
-
-		// Create key for this time slot
-		key := fmt.Sprintf("%02d:%02d", hour, minute)
-
-		// Add goal to this time slot
-		if slot, exists := slotMap[key]; exists {
-			slot.goals = append(slot.goals, goal.Slug)
-		} else {
-			slotMap[key] = &timeSlot{
-				hour:   hour,
-				minute: minute,
-				goals:  []string{goal.Slug},
-			}
-		}
-	}
-
-	// Convert map to sorted slice
-	var slots []timeSlot
-	for _, slot := range slotMap {
-		slots = append(slots, *slot)
-	}
-
-	// Sort by time (hour, then minute)
-	sort.Slice(slots, func(i, j int) bool {
-		if slots[i].hour != slots[j].hour {
-			return slots[i].hour < slots[j].hour
-		}
-		return slots[i].minute < slots[j].minute
-	})
-
-	return slots
-}
-
-// displayHourlyDensity displays a compact bar chart showing goals per hour
-func displayHourlyDensity(hourCounts []int) {
-	const (
-		hoursPerDay    = 24
-		charsPerHour   = 3 // Each hour uses 3 chars: 2 for display (bar/label) + 1 space
-		densityLineLen = hoursPerDay * charsPerHour
-	)
-
-	fmt.Println("HOURLY DENSITY")
-
-	// Find max count for scaling
-	maxCount := 0
-	for _, count := range hourCounts {
-		if count > maxCount {
-			maxCount = count
-		}
-	}
-
-	// If no goals, just display empty chart
-	if maxCount == 0 {
-		fmt.Println("No goals scheduled.")
-		return
-	}
-
-	// Define bar characters (from empty to full)
-	bars := []rune{' ', '▁', '▂', '▃', '▄', '▅', '▆', '▇', '█'}
-
-	// Build the bar chart line
-	var barLine strings.Builder
-
-	for hour := 0; hour < hoursPerDay; hour++ {
-		count := hourCounts[hour]
-		var bar rune
-		if count == 0 {
-			bar = bars[0]
-		} else {
-			// Scale to bar height (1-8)
-			barIndex := (count * (len(bars) - 1)) / maxCount
-			if barIndex == 0 && count > 0 {
-				barIndex = 1 // Ensure at least one bar level for any count
-			}
-			bar = bars[barIndex]
-		}
-		// Write the bar twice for 2-char width per hour, plus a space
-		barLine.WriteRune(bar)
-		barLine.WriteRune(bar)
-		barLine.WriteRune(' ')
-	}
-
-	fmt.Println(barLine.String())
-
-	// Build hour labels (show all 24 hours with spacing)
-	// Use color to de-emphasize hours with no counts (if colors enabled)
-	colorProfile := lipgloss.ColorProfile()
-	dimStyle := lipgloss.NewStyle().Foreground(lipgloss.Color("8"))
-	var labelLine strings.Builder
-
-	for hour := 0; hour < hoursPerDay; hour++ {
-		label := fmt.Sprintf("%02d", hour)
-		if hourCounts[hour] == 0 && colorProfile != termenv.Ascii {
-			// Dim hours with no counts using gray color
-			labelLine.WriteString(dimStyle.Render(label) + " ")
-		} else {
-			labelLine.WriteString(label + " ")
-		}
-	}
-	fmt.Println(labelLine.String())
-
-	// Build axis line with markers for each hour
-	axisRunes := make([]rune, densityLineLen)
-	for i := range axisRunes {
-		axisRunes[i] = ' '
-	}
-
-	firstPos, lastPos := -1, -1
-	for hour := 0; hour < hoursPerDay; hour++ {
-		pos := hour * charsPerHour
-		if pos >= len(axisRunes) {
-			continue
-		}
-		// Use ┼ if hour has counts (extends down), ┴ if no counts (only extends up)
-		if hourCounts[hour] > 0 {
-			axisRunes[pos] = '┼'
-		} else {
-			axisRunes[pos] = '┴'
-		}
-		if firstPos == -1 {
-			firstPos = pos
-		}
-		lastPos = pos
-	}
-
-	// Draw horizontal line segments between first and last marker
-	if firstPos != -1 && lastPos != -1 && firstPos <= lastPos {
-		for i := firstPos + 1; i < lastPos; i++ {
-			if axisRunes[i] == ' ' {
-				axisRunes[i] = '─'
-			}
-		}
-	}
-
-	fmt.Println(string(axisRunes))
-
-	// Build count labels (show counts for all hours with goals)
-	countLine := make([]rune, densityLineLen)
-	for i := range countLine {
-		countLine[i] = ' '
-	}
-
-	for hour := 0; hour < hoursPerDay; hour++ {
-		count := hourCounts[hour]
-		if count > 0 {
-			var label string
-			if count > 99 {
-				label = " \u221e" // Use infinity symbol for 99+, with leading space
-			} else {
-				label = fmt.Sprintf("%-2d", count) // Left-align in 2-char space
-			}
-			pos := hour * charsPerHour
-			// Write the label runes into the count line, guarding against overflow
-			for i, r := range label {
-				if pos+i >= len(countLine) {
-					break
-				}
-				countLine[pos+i] = r
-			}
-		}
-	}
-	fmt.Println(string(countLine))
-	fmt.Println()
-}
-
-// displayTimeline displays a vertical timeline listing all goals grouped by deadline time
-func displayTimeline(slots []timeSlot) {
-	fmt.Println("TIMELINE")
-	fmt.Println("────────────────────────────────────────────────")
-
-	// Define styles for timeline elements (disabled if --no-color)
-	colorProfile := lipgloss.ColorProfile()
-	timeStyle := lipgloss.NewStyle()
-	treeStyle := lipgloss.NewStyle()
-	if colorProfile != termenv.Ascii {
-		timeStyle = timeStyle.Foreground(lipgloss.Color("6")) // Cyan for time labels
-		treeStyle = treeStyle.Foreground(lipgloss.Color("8")) // Gray for tree structure (├─ and │)
-	}
-
-	for _, slot := range slots {
-		timeStr := fmt.Sprintf("%02d:%02d", slot.hour, slot.minute)
-		goalsStr := strings.Join(slot.goals, ", ")
-
-		// Build the full line and wrap it to terminal width, indenting wrapped lines
-		// Color the time and tree separately using lipgloss
-		prefix := timeStyle.Render(timeStr) + " " + treeStyle.Render("├─") + " "
-		// Visual width of prefix: "HH:MM ├─ " = 9 characters (ANSI codes have zero width)
-		const prefixVisualWidth = 9
-
-		// Determine terminal width; fallback to 80 if unavailable
-		width := 80
-		fd := uintptr(os.Stdout.Fd())
-		if term.IsTerminal(fd) {
-			if w, _, err := term.GetSize(fd); err == nil && w > 0 {
-				width = w
-			}
-		}
-
-		// Simple wrapping: break on commas before exceeding width
-		available := width
-		if prefixVisualWidth < available {
-			available -= prefixVisualWidth
-		} else {
-			available = 10 // minimal safety width
-		}
-
-		// Split on commas to get individual goals
-		goals := strings.Split(goalsStr, ", ")
-		var line strings.Builder
-		line.WriteString(prefix)
-		current := 0
-		for _, goal := range goals {
-			// Determine separator
-			sep := ""
-			if current > 0 {
-				// Add comma+space before this goal if there's already content on this line
-				sep = ", "
-			}
-			chunk := sep + goal
-			if current+len(chunk) > available && current > 0 {
-				// Always add trailing comma when wrapping (more content follows)
-				line.WriteString(",")
-				fmt.Println(line.String())
-				// start new wrapped line with indent matching prefix width
-				// Use vertical line to show continuation
-				line.Reset()
-				// Time column width (5 chars) + space + vertical continuation (styled)
-				line.WriteString("      " + treeStyle.Render("│") + "  ")
-				// Add the goal that didn't fit
-				line.WriteString(goal)
-				current = len(goal)
-				continue
-			}
-			line.WriteString(chunk)
-			current += len(chunk)
-		}
-		fmt.Println(line.String())
-	}
-}
-
-// handleUncleCommand instantly derails a goal that is in the red.
-func handleUncleCommand() {
-	uncleFlags := flag.NewFlagSet("uncle", flag.ContinueOnError)
-	uncleFlags.Usage = func() {
-		fmt.Fprintln(os.Stderr, "Usage: buzz uncle [-y|--yes] <goalslug>")
-	}
-	yes := uncleFlags.Bool("yes", false, "Skip the confirmation prompt")
-	yesShort := uncleFlags.Bool("y", false, "Skip the confirmation prompt (shorthand)")
-	if err := uncleFlags.Parse(os.Args[2:]); err != nil {
-		if errors.Is(err, flag.ErrHelp) {
-			uncleFlags.Usage()
-			return
-		}
-		fmt.Fprintf(os.Stderr, "Error parsing flags: %s\n", err)
-		fmt.Fprintln(os.Stderr, "Usage: buzz uncle [-y|--yes] <goalslug>")
-		os.Exit(2)
-	}
-
-	args := uncleFlags.Args()
-	if len(args) != 1 {
-		if len(args) == 0 {
-			fmt.Fprintln(os.Stderr, "Error: Missing required argument")
-		} else {
-			fmt.Fprintf(os.Stderr, "Error: Too many arguments: %v\n", args[1:])
-		}
-		fmt.Fprintln(os.Stderr, "Usage: buzz uncle [-y|--yes] <goalslug>")
-		os.Exit(1)
-	}
-
-	goalSlug := args[0]
-	skipConfirm := *yes || *yesShort
-
-	if !ConfigExists() {
-		fmt.Fprintln(os.Stderr, "Error: No configuration found. Please run 'buzz' first to authenticate.")
-		os.Exit(1)
-	}
-
-	config, err := LoadConfig()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: Failed to load config: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	client := NewHTTPClient(config)
-
-	if !skipConfirm {
-		fmt.Printf("Call uncle on %s? This will instantly derail the goal and charge the pledge. [y/N] ", goalSlug)
-		var response string
-		fmt.Scanln(&response)
-		response = strings.TrimSpace(strings.ToLower(response))
-		if response != "y" && response != "yes" {
-			fmt.Println("Cancelled.")
-			return
-		}
-	}
-
-	goal, err := client.CallUncle(goalSlug)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: Failed to call uncle: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	fmt.Printf("Called uncle on %s. The goal has been derailed.\n", goal.Slug)
-
-	fmt.Print(getUpdateMessage())
-}
